Amadeus/requirements.txt
discord.py[voice]==2.6.3
python-dotenv==1.0.1
aiosqlite==0.20.0
humanize==4.10.0
matplotlib==3.9.2
psutil==6.0.0
flask==2.3.3

